Support Scala 3.x Migrate from automatic to semi-automatic circe codec derivation to work around incomplete support.
