+++ .github/workflows/ci.yml
@@ -23,7 +23,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest]
-        scala: [2.13.8]
+        scala: [2.13.8, 3.1.2]
         java: [graal_21.3.0@11, graal_21.3.0@17]
         platform: [jvm, js]
     runs-on: ${{ matrix.os }}
